Add Validation Logic to api* Models and Implement Validator in apiController
This PR addresses a bug causing DB insert failures to occur because of un-cleansed input. Validation logic lives within the models, and is passed into a Validator in the concerned controller.

Both Jobs are now completing without any failures.

TODO: Look into implementing a BaseModel which provides validation functionality for the data integrity of the model itself, rather than of the data being used to hydrate it.